services/market-data-collector/requirements.txt
websocket-client==1.7.0
kafka-python==2.0.2
python-dotenv==1.0.1
